+++ MainMethodTemplate.cs
@@ -1,9 +1,11 @@
+static Settings Settings => GL._Settings;
+
 static void Main(string[] args)
 {
     Console.WriteLine("Hello, World!");
     GL._Settings = Settings.Args(typeof(Program).Namespace!.ToString(), args, false, false, false, null);
 
-    string path = GL._Settings.SetDefaultOutputPaths("DefaultOutputPath", "DefaultOutputName");
+    string path = GL._Settings.SetDefaultOutputPaths("Reports", "DefaultOutputName");